refactor(tenants): extract SettingsTab component
